pyproject.toml
[tool.poetry]
name = "wow-fisher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
evaluate = "^0.4.3"
fire = "^0.7.0"
ipykernel = "^6.29.5"
loguru = "^0.7.2"
pandas = "^2.2.3"
pillow = "^11.0.0"
pyaudiowpatch = "^0.2.12.6"
pydantic = "^2.9.2"
pynput = "^1.7.7"
python = "^3.10"
pyyaml = "^6.0.2"
scikit-learn = "^1.5.2"
soundfile = "^0.12.1"
torch = {version = "^2.5.1+cu124", source = "pytorch"}
torchaudio = {version = "^2.5.1+cu124", source = "pytorch"}
torchvision = {version = "^0.20.1+cu124", source = "pytorch"}
transformers = {extras = ["torch"], version = "^4.46.3"}
ultralytics = "^8.2.76"
pyaudio = "^0.2.14"


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.cloud.tencent.com/pypi/simple"
priority = "primary"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
